# microservices/oa-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: '0.1.0'
  title: Open Affect API
  description: API to record emotions
  termsOfService: none
  contact:
    name: Open Affect API team
    email: info@openaffect.io
    url: http://openaffect.io
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: 192.168.99.100
basePath: /api
schemes:
  - http
paths:
  /measures:
    post:
      description: report a new measure (used by a sensor)
      operationId: reportMeasure
      consumes:
        - application/json
      parameters:
        - name: measure
          in: body
          description: measure reported by the sensor
          required: true
          schema:
            $ref: '#/definitions/measure'
      responses:
        '201':
          description: no body
          schema:
            type: object
    get:
      description: list reported measures
      operationId: listMeasures
      produces:
        - application/json
      responses:
        '200':
          description: list of measures
          schema:
            type: array
            items:
              $ref: '#/definitions/measure'
      parameters:
        - name: triggerHref
          in: query
          type: string
          description: the trigger href field to look for
          required: false
        - name: fromDate
          in: query
          type: string
          format: date-time
          description: the inclusive date from which the measures are collected (only works if the triggerHref parameter was given)
          required: false
        - name: toDate
          in: query
          type: string
          format: date-time
          description: the inclusive date to which the measures are collected (only works if the triggerHref parameter was given)
          required: false
    put:
      description: edit a measure by the given trigger's href.
      operationId: editMeasure
      consumes:
        - application/json
      parameters:
        - name: triggerHref
          in: query
          type: string
          description: the trigger's href of the measure
          required: true
        - name: editedMeasure
          in: body
          required: true
          schema:
            $ref: '#/definitions/measure'
      responses:
        '200':
          description: edited
          schema:
            type: object
    delete:
      description: delete measures related to one trigger
      operationId: deleteMeasuresByTrigger
      parameters:
        - name: triggerHref
          in: query
          type: string
          description: the trigger's href of the measures to delete
          required: true
      responses:
        '204':
          description: deleted
definitions:
  measure:
    type: object
    required:
      - timestamp
      - emotion
      - subject
      - trigger
      - sensor
    properties:
      timestamp:
        type: string
        format: date-time
      emotion:
        $ref: '#/definitions/emotion'
      subject:
        $ref: '#/definitions/resource'
      trigger:
        $ref: '#/definitions/resource'
      sensor:
        $ref: '#/definitions/resource'
  resource:
      type: object
      required:
        - href
        - type
        - properties
      properties:
        href:
          type: string
        type:
          type: string
        properties:
          type: object
          additionalProperties:
            type: object
  emotion:
    type: object
    properties:
      intensity:
        type: number
        format: double
      category:
        type: string
      properties:
          type: object
          additionalProperties:
            type: object